\name{bmhyd-package}
\alias{bmhyd-package}
\alias{bmhyd}
\docType{package}
\title{
Brownian motion model for phylogenetic comparative analysis on hybridization.}
\description{
Investigation the hybrid effect for evolution along a phylognenetic network.
}
\details{
\tabular{ll}{
Package: \tab bmhyd\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-01-21\cr
License: \tab  GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Tony Jhwueng <djhwueng@nimbios.org>


Maintainer:Tony Jhwueng <djhwueng@nimbios.org>
}
\references{
A new comparative analysis for investigating rate
of trait evolution using phylogenetic networks. manuscript submmitted in Evolution. }
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
